added debug collision drawer + fixes

+++ src/Components/BoxColliderComponent.h
@@ -5,11 +5,11 @@
 #include "glm/glm.hpp"
 
 struct BoxColliderComponent {
-  BoxColliderComponent(int width = 0, int height = 0, glm::vec2 offset = {})
+  BoxColliderComponent(double width = 0., double height = 0., glm::vec2 offset = {})
       : width(width), height(height), offset(offset) {}
 
-  int width{};
-  int height{};
+  double width{};
+  double height{};
   glm::vec2 offset{};
 
   bool GetIsColliding() { return _isColliding; }
